+++ components/SelectMoodMeterTab.tsx
